Finido the lossless compresso

+++ src/predictor/localsum.c
@@ -6,7 +6,7 @@
 	Chapter 4.4 in CCSDS 123 Issue 2
 */
 
-void wideNeighborLocalSum(int * sample, int * localsum, int x, int y, int z, struct arguments * parameters) {
+void wideNeighborLocalSum(unsigned long * sample, long * localsum, int x, int y, int z, struct arguments * parameters) {
 	if (x+y != 0) {
 		if (y == 0 && x > 0){
 			localsum[offset(x,y,z,parameters)] = 4 * sample[offset(x-1,y,z, parameters)];
@@ -20,7 +20,7 @@ void wideNeighborLocalSum(int * sample, int * localsum, int x, int y, int z, str
 	}
 }
 
-void narrowNeighborLocalSum(int * sample, int * localsum, int smid, int x, int y, int z, struct arguments * parameters) {
+void narrowNeighborLocalSum(unsigned long * sample, long * localsum, int smid, int x, int y, int z, struct arguments * parameters) {
 	if(!(x == 0 && y == 0)) {
 		if (y > 0 && x > 0 && x < parameters->xSize-1) {
 			localsum[offset(x,y,z,parameters)] = sample[offset(x-1,y-1,z, parameters)] +  (2 * sample[offset(x,y-1,z, parameters)]) + sample[offset(x+1,y-1,z, parameters)];
@@ -38,7 +38,7 @@ void narrowNeighborLocalSum(int * sample, int * localsum, int smid, int x, int y
 	}
 }
 
-void wideColumnLocalSum(int * sample, int * localsum, int x, int y, int z, struct arguments * parameters) {
+void wideColumnLocalSum(unsigned long * sample, long * localsum, int x, int y, int z, struct arguments * parameters) {
 	if(!(x == 0 && y == 0)) {
 		if(y > 0) {
 			localsum[offset(x,y,z,parameters)] = 4 * sample[offset(x,y-1,z, parameters)];
@@ -50,7 +50,7 @@ void wideColumnLocalSum(int * sample, int * localsum, int x, int y, int z, struc
 	}
 }
 
-void narrowColumnLocalSum(int * sample, int * localsum, int smid, int x, int y, int z, struct arguments * parameters) {
+void narrowColumnLocalSum(unsigned long * sample, long * localsum, int smid, int x, int y, int z, struct arguments * parameters) {
 	if(!(x == 0 && y == 0)) {
 		if (y > 0) {
 			localsum[offset(x,y,z,parameters)] = 4 * sample[offset(x,y-1,z, parameters)];
