+++ src/utils/include/utilities.h
@@ -1,3 +1,3 @@
 #pragma once
 #include "../../cliparser/include/cli.h"
-int readIntSamples(struct arguments * parameters, char fileName[128], unsigned int * samples);
+int readIntSamples(struct arguments * parameters, char fileName[128], unsigned long * samples);
